Added support for ZIP64 archives.

+++ zipdefs.h
@@ -33,8 +33,8 @@ struct localheader {
     uint16_t last_mod_time;
     uint16_t last_mod_date;
     uint8_t crc32[4];
-    uint32_t compressed_size;
-    uint32_t uncompressed_size;
+    uint64_t compressed_size; // On disk header format is 32 bits, but this is 64 bits to be able to store zip64 offsets, too.
+    uint64_t uncompressed_size;
     std::string fname;
     std::string extra;
 };
